# Optimize `Article.published` scope to use an index