--- a/app/models/article.rb
+++ b/app/models/article.rb
@@ -110,6 +110,11 @@ class Article < ApplicationRecord
   serialize :cached_user
   serialize :cached_organization
 
+  # [@jgaskins] We use an index on `published`, but since it's a boolean value
+  #   the Postgres query planner often skips it due to lack of diversity of the
+  #   data in the column. However, since `published_at` is a *very* diverse
+  #   column and can scope down the result set significantly, the query planner
+  #   can make heavy use of it.
   scope :published, -> {
     self
       .where(published: true)