--- a/spec/factories/articles.rb
+++ b/spec/factories/articles.rb
@@ -5,7 +5,7 @@ FactoryBot.define do
     transient do
       title { generate :title }
       published { true }
-      published_at { Time.now }
+      published_at { Time.current }
       date { "01/01/2015" }
       tags { "javascript, html, css" }
       canonical_url { Faker::Internet.url }