--- a/spec/factories/articles.rb
+++ b/spec/factories/articles.rb
@@ -2,10 +2,11 @@ FactoryBot.define do
   sequence(:title) { |n| "#{Faker::Book.title}#{n}" }
 
   factory :article do
+    published_at { Time.current }
+
     transient do
       title { generate :title }
       published { true }
-      published_at { Time.current }
       date { "01/01/2015" }
       tags { "javascript, html, css" }
       canonical_url { Faker::Internet.url }