@jgaskins 2021-03-23
@vaidehijoshi @rhymes @joshpuetz Can y'all just double-check the updates to the specs I made after you reviewed? It was mostly making sure `published_at` was set when `published = true` and making sure it wasn't when it's not, which seems to be what we're doing already in files like [`spec/requests/api/v0/articles_spec.rb`](https://github.com/forem/forem/blob/master/spec/requests/api/v0/articles_spec.rb#L506), but also I wanted to make sure before I merge that I'm not just changing specs to make them pass. 😄 

If it makes sense, I'd like to go ahead and merge ASAP so we can start collecting the new perf metrics!